requirements.txt
uvicorn
"fastapi[standard]"
openai
python-gitlab
loguru
munch